Remove runtime user canvas detection
The optional canvas component would require a second render in order to
work. In the first pass, we'd check whether the user supplied a canvas
and then, in the second pass, decide whether to render the default
canvas.

This is, of course, not ideal.

Judging by issue #99, someone has finally found an edge case, where
you can end up fetching a DOM element from the first pass and watch it
suddenly disappear.