Replace `ObjectProxy` and `PromiseProxyMixin`
These remnants of Ember’s past help working with promises in templates.
Together they act as a proxy that tells Ember to rerender things
when the promise resolves.

The main problem is that they’ll be removed at some point, and probably
sooner than later. We can write our own replacement, but it requires
`Proxy`, which isn’t IE11-friendly. So for now we’ll have to run both
versions side by side, but we’re ready to seamlessly drop the old
version at any time.