Reimplement the addon system using Embroider macros
This is a decent option actually! Not sure it’s it in the spirit of
macros though. The AST transformer doesn’t seem to work well with
Embroider. It keeps trying to delete the extra addons from the build.
Perhaps it’s analyzing the template before we run the transformation
and decides it can delete the addon?