Detect Embroider and disable tree filtering
It’s difficult to figure out how the classic `treeForAddon` and
`setupPreprocessor` methods fit into Embroider’s build pipeline,
particularly when it comes to the order in which things get executed.
And since Embroider does its own analysis and eventual treeshaking, it
doesn’t make sense to even try to filter stuff out. I think Embroider
doesn’t expect stuff to be deleted from the trees and be used in the
template, so it errors out.