Don't error out when setting an undefined option
We were setting `center` to `undefined` when drawing a map without
coordinates. That actually produced a reasonable error message because
the setter returned `false` when trying to delete the non-existent
`center` key from the set cache.

We now return `true` from the setter so that it doesn't error out.
This also means that `center` is now an optional argument.

Fixes #120.