Fix issue where directions were re-fetched when resolved
`ember-concurrency` tracks its internal properties, so it ends up
triggering the effect a second time once it resolves. I guess we
could "changeset" the options to avoid this, but there's more. Because
it runs in the same computation as this effect, you can't even
check`isRunning` without triggering an error from Glimmer. That's not
particularly great, and I guess the solution might have to happen
upstream (scheduling or track/untrack frames). Let's see what comes out
of Glimmer's effect system and revisit.