Refactor AST plugins to match new plugin format
There isn't any documentation on how to write these AST plugins. Until
now, I would base the design on whatever rjwblue was doing, because,
you know, core-team and everything. With Ember@4.0 the class-based
design he uses doesn't work. The plugin now HAS to be a function that
returns an object with the visitor (see here:
https://github.com/glimmerjs/glimmer-vm/blob/ef92c497c3d921af22b38a54eac1445c204c3370/packages/%40glimmer/syntax/lib/parser/tokenizer-event-handlers.ts#L318-L325
)