Update the coordinates when `lat` and `lng` change
`lat` and `lng` are convenience options that need to be converted to
either `center` or `position`, depending on the map component. However,
we don’t want to overwrite either those options when they’re provided,
so we check before setting them. With the Octane rewrite, option
overwrites are persisted between updates. Now, because of those checks,
we might fail to update the coordinates if we’ve already set them
once.

Fixes #150.