Fix tests against ember `>=4`
The latest beta of Ember v4 now includes ember-auto-import `>=2` as a
dependency, forcing the consuming application to include it as well.
As a side-effect of that, we can’t run our test suite against `v4`
unless we also add ember-auto-import to our dependencies, again forcing
apps downstream to do the same.

We don’t use any of the features of auto-import, so it feels rather
silly dropping a major version over this. Luckily, we can use ember-try
scenarios to work around this!