Remove the in-repo test addon from package.json
This in-repo addon is used for testing and is dynamically removed from
the addon tree when not needed. Embroider is now failing to rebuild the
app on changes because it's expecting files from this addon to be
present. I've removed the addon from package.json and disabled the test
for now. We'll have to migrate to the v2 addon format at some point and
reintegrate the test.